Updated projection features
Adjusted map extent so polygons extending furthest south in the middle of the map will be shown. This was an issue with the orthographic projection being used.
Added state labels for maps which will not have city labels and doesn't show the majority of the US.
Fixed bug where variable wasn't set for days 3-8.