Title and Label Updates
Time zones are only found at locations on land, which removes bugs associated with daylight savings time.
The legend is no longer labelled with words. Instead, numbers are used. The TSTM label is lightning emojis so it isn't much longer than the rest of the labels.
The change to the legend reflects research by Sean Ernst and Kenzie Krocak suggesting the public understands the numbers, but the words cause confusion and putting both numbers and words don't help. People default back to the confusing words. More is not better.